implemented nav bar, cleaned color css
