+++ package.json
@@ -1,37 +1,38 @@
 {
-  "name": "movie-rating",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "homepage": "https://Chungywungy.github.io/movie-rating",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d dist"
-  },
-  "dependencies": {
-    "axios": "^1.8.4",
-    "firebase": "^11.6.0",
-    "lodash.debounce": "^4.0.8",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.21.0",
-    "@tailwindcss/postcss": "^4.1.3",
-    "@types/react": "^19.0.10",
-    "@types/react-dom": "^19.0.4",
-    "@vitejs/plugin-react": "^4.3.4",
-    "autoprefixer": "^10.4.21",
-    "eslint": "^9.21.0",
-    "eslint-plugin-react-hooks": "^5.1.0",
-    "eslint-plugin-react-refresh": "^0.4.19",
-    "gh-pages": "^6.3.0",
-    "globals": "^15.15.0",
-    "postcss": "^8.5.3",
-    "tailwindcss": "^4.1.3",
-    "vite": "^6.2.0"
-  }
+    "name": "movie-rating",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "homepage": "https://Chungywungy.github.io/movie-rating",
+    "scripts": {
+        "dev": "vite",
+        "build": "vite build",
+        "preview": "vite preview",
+        "predeploy": "npm run build",
+        "deploy": "gh-pages -d dist"
+    },
+    "dependencies": {
+        "axios": "^1.8.4",
+        "firebase": "^11.6.0",
+        "lodash.debounce": "^4.0.8",
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0"
+    },
+    "devDependencies": {
+        "@eslint/js": "^9.21.0",
+        "@tailwindcss/postcss": "^4.1.3",
+        "@types/react": "^19.0.10",
+        "@types/react-dom": "^19.0.4",
+        "@vitejs/plugin-react": "^4.3.4",
+        "autoprefixer": "^10.4.21",
+        "eslint": "^9.21.0",
+        "eslint-plugin-react-hooks": "^5.1.0",
+        "eslint-plugin-react-refresh": "^0.4.19",
+        "gh-pages": "^6.3.0",
+        "globals": "^15.15.0",
+        "postcss": "^8.5.3",
+        "react-router-dom": "^7.6.0",
+        "tailwindcss": "^4.1.3",
+        "vite": "^6.2.0"
+    }
 }
