--- src/MovieSearch.css
@@ -1,163 +1,0 @@
-/* Base layout */
-html, body {
-  margin: 0;
-  padding: 0;
-  background-color: #0d1117;
-  color: #f0f6fc;
-  font-family: sans-serif;
-  height: 100%;
-}
-.app {
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    font-family: sans-serif;
-    background: #0d1117; /* dark navy */
-    color: #f0f6fc;       /* soft blue text */
-  }
-  
-  /* Search form */
-  form {
-    margin-bottom: 20px;
-  }
-  
-  input {
-    padding: 8px;
-    width: 100%;
-    box-sizing: border-box;
-    margin-bottom: 10px;
-    background-color: #161b22;
-    color: #f0f6fc;
-    border: 1px solid #30363d;
-    border-radius: 4px;
-  }
-  
-  input::placeholder {
-    color: #8b949e;
-  }
-  
-  button {
-    padding: 8px 12px;
-    border: none;
-    cursor: pointer;
-    background-color: #072551;
-    color: #fff;
-    border-radius: 4px;
-    transition: background-color 0.2s ease-in-out;
-  }
-  
-  button:hover {
-    background-color: #388bfd;
-  }
-  
-  /* Movie card container */
-  .movie-card {
-    background-color: #161b22;
-    border: 1px solid #30363d;
-    padding: 15px;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0,0,0,0.2);
-  }
-  
-  /* Poster image */
-  img {
-    max-width: 100%;
-    height: auto;
-    border-radius: 5px;
-    margin: 10px 0;
-  }
-  
-  .rating-buttons {
-    margin-top: 10px;
-  }
-  
-  .rating-buttons button {
-    padding: 6px 10px;
-    border-radius: 50%;
-    margin-right: 5px;
-    border: none;
-    font-weight: bold;
-  }
-  
-  .selected-rating {
-    background-color: #1f6feb;
-    color: #fff;
-  }
-  
-  .unselected-rating {
-    background-color: #21262d;
-    color: #f0f6fc;
-  }
-
-  .rated-movies {
-  margin-top: 40px;
-  padding: 15px;
-  background-color: #161b22;
-  border: 1px solid #30363d;
-  border-radius: 8px;
-}
-
-.rated-movies h3 {
-  color: #ffd700; /* gold */
-}
-
-.rated-movies ul {
-  list-style: none;
-  padding: 0;
-}
-
-.rated-movies li {
-  padding: 5px 0;
-  border-bottom: 1px solid #30363d;
-}
-  
-.rated-movies {
-  margin-top: 40px;
-  padding: 15px;
-  background-color: #161b22;
-  border: 1px solid #30363d;
-  border-radius: 8px;
-}
-
-.rated-movies h3 {
-  margin-bottom: 10px;
-  color: #58a6ff;
-}
-
-.rated-movies ul {
-  list-style: none;
-  padding: 0;
-}
-
-.rated-movies li {
-  padding: 5px 0;
-  border-bottom: 1px solid #30363d;
-}
-
-.dropdown {
-  list-style: none;
-  margin-top: 8px;
-  padding: 0;
-  border: 1px solid #30363d;
-  background-color: #293c56;
-  border-radius: 6px;
-  max-height: 200px;
-  overflow-y: auto;
-  position: absolute;
-  z-index: 10;
-  width: 100%;
-}
-
-.dropdown li {
-  padding: 10px;
-  cursor: pointer;
-  border-bottom: 1px solid #30363d;
-}
-
-.dropdown li:last-child {
-  border-bottom: none;
-}
-
-.dropdown li:hover {
-  background-color: #21262d;
-}
