+++ src/colors.css
@@ -1,0 +1,7 @@
+:root {
+    --primary-color: #ed254e;
+    --secondary-color: #f9dc5c;
+    --background-color: #f4fffd;
+    --accent-color: #011936;
+    --text-color: #465362;
+}
