+++ src/components/Logo/Logo.css
@@ -1,0 +1,6 @@
+.logo {
+    color: var(--background-color);
+    margin: auto;
+    padding-left: 1rem;
+    padding-right: 1rem;
+}
