+++ src/components/NavBarButton/NavBarButton.jsx
@@ -1,0 +1,21 @@
+import './NavBarButton.css';
+import { useNavigate } from 'react-router-dom';
+
+export default function NavBarButton({ className, to, children }) {
+    const navigate = useNavigate();
+
+    function handleClick() {
+        navigate(to);
+    }
+
+    return (
+        <>
+            <button
+                className={`nav-bar-button ${className}`}
+                onClick={handleClick}
+            >
+                {children}
+            </button>
+        </>
+    );
+}
