--- src/index.css
@@ -1,13 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-img {
-    border: none !important;
-    outline: none !important;
-    background: transparent !important;
-  }
-
-  ::-webkit-scrollbar {
-    width: 0px;
-    background: transparent;
-  }
